Bot/requirements.txt
discord.py==2.6.3
PyYAML==6.0.2